+++ crates/ruma-client-api/src/r0/contact/get_contacts.rs
@@ -1,7 +1,7 @@
 //! [GET /_matrix/client/r0/account/3pid](https://matrix.org/docs/spec/client_server/r0.6.0#get-matrix-client-r0-account-3pid)
 
 use ruma_api::ruma_api;
-pub use ruma_common::thirdparty::{ThirdPartyIdentifier, ThirdPartyIdentifierInit};
+use ruma_common::thirdparty::ThirdPartyIdentifier;
 
 ruma_api! {
     metadata: {