+++ crates/ruma-client-api/src/r0/to_device.rs
@@ -1,5 +1,3 @@
 //! Endpoints for client devices to exchange information not persisted in room DAG.
 
-pub use ruma_common::to_device::DeviceIdOrAllDevices;
-
 pub mod send_event_to_device;